pyproject.toml
[project]
name = "pichia-kb"
version = "0.1.0"
description = "Pichia pastoris fermentation knowledge base and Q&A system"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "google-genai>=1.73.1",
    "hatchling>=1.29.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.13.3",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
    "streamlit>=1.56.0",
    "typer>=0.24.2",
]

[project.scripts]
pichia-kb = "pichia_kb.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pichia_kb"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
